refactor: viewAllPage 파일명 변경
